Added Script
If the command 'python ResampleScript.py is run with either 1 or 2 dicom directorys or .mnc files with a third RT file. If the input is a dicom directory the script also checks if the RT was defined on this image set. If the input is a dicom directory its convertet to .mnc format. 
Then the .mnc file is run with the RT file and the RT file then becomes 'filled' and is outputtet as an .mnc file. 
Now its resampled as the input with the option of resampling the input as the RT.mnc file.